==> 1_twitter_api_01.py <==
# -*- coding: utf-8 -*-
"""1_twitter_api_01.ipynb

Automatically generated by Colaboratory.

Original file is located at
    https://colab.research.google.com/drive/1Az-SLYIgEWmwEapbILgVbHpSag4_C_sp
"""

# twitter api doc API reference index
# https://developer.twitter.com/en/docs/api-reference-index

import platform
import requests

# 12 nba stars twitter 
nba_star = ["KDTrey5", "JHarden13", "russwest44", "CP3", "Giannis_An34", 
       "Dame_Lillard", "TheTraeYoung", "DevinBook", "ZachLaVine",
       "StephenCurry30", "KlayThompson", "DeMar_DeRozan"]

token = "AAAAAAAAAAAAAAAAAAAAANTvWAEAAAAAxV1p50rhTxTNIwscXWaR6qcDoNg%3D7CdaQpbNovUrSLL5V3dyZGx62zLDkBeW5RzGvqBn77rHfTImE4"


def get_user_info(username):
  url = f"https://api.twitter.com/2/users/by/username/{username}"
  headers = {'Authorization': f'Bearer {token}'}
  params = {"user.fields": "description,url,location"} # 
  response = requests.get(url, headers=headers, params=params)
  return response.json()

username = "KDTrey5"
get_user_info(username)

import pandas as pd 
user_data = {"user_id": [], "name": [], "username": [],
        "description": [], "url": [], "location": []}
             
for i in nba_star:
  x = get_user_info(i)["data"]
  user_data["user_id"].append(x["id"])
  user_data["name"].append(x["name"])
  user_data["username"].append(x["username"])

  user_data["description"].append(x["description"] if "description" in x else None)
  user_data["url"].append(x["url"] if "url" in x else None)
  user_data["location"].append(x["location"] if "location" in x else None)

user_df = pd.DataFrame(user_data)
user_df

from google.colab import files
user_df.to_csv("user_info.csv")

files.download("user_info.csv")

"""# Get followers. Maybe later"""

# doc max results 100 need turn pages.
# https://developer.twitter.com/en/docs/twitter-api/users/follows/api-reference/get-users-id-followers

# def get_followers(user_id):
#   url = f"https://api.twitter.com/2/users/{user_id}/followers"
#   headers = {'Authorization': f'Bearer {token}'}
#   params = {"max_results": 100}
#   response = requests.get(url, headers=headers)
#   return response.json()

# get_followers("43658360")

"""# Get following"""

# doc max results 100  
# https://developer.twitter.com/en/docs/twitter-api/users/follows/api-reference/get-users-id-followers

def get_following(user_id):
  url = f"https://api.twitter.com/2/users/{user_id}/following"
  headers = {'Authorization': f'Bearer {token}'}
  params = {"max_results": 100}
  response = requests.get(url, headers=headers, params=params)
  return response.json()

# get_following("35936474")

following_dic = {"user_id": [], "following_user_id": [], "following_name": [], "following_username": []}
for i in user_df["user_id"]:
  for j in get_following(i)["data"]:
    following_dic["user_id"].append(i)
    following_dic["following_user_id"].append(j["id"])
    following_dic["following_name"].append(j["name"])
    following_dic["following_username"].append(j["username"])


following_df = pd.DataFrame(following_dic)
following_df

from google.colab import files
user_df.to_csv("user_with_follwing.csv", index=False)

files.download("user_with_follwing.csv")

"""# Get 100 tweets for each user"""

# get 100 tweets
def get_tweets(user_id):
  url = f"https://api.twitter.com/2/users/{user_id}/tweets"
  headers = {'Authorization': f'Bearer {token}'}
  params = {"max_results": 100}
  response = requests.get(url, headers=headers, params=params)
  return response.json()

get_tweets("35936474")

tweets_dic = {"user_id": [], "tweet_id": [], "text": []}
for i in user_df["user_id"]:
  for j in get_tweets(i)["data"]:
    tweets_dic["user_id"].append(i)
    tweets_dic["tweet_id"].append(j["id"])
    tweets_dic["text"].append(j["text"])
    
tweets_df = pd.DataFrame(tweets_dic)
tweets_df

from google.colab import files
tweets_df.to_csv("user_tweets.csv", index=False)

files.download("user_tweets.csv")
